third_party/blink/web_tests/platform/generic/fast/replaced/preferred-widths-expected.txt

Test the effect of percentages widths on the preferred widths of replaced elements.

PASS
PASS
FAIL:
Expected 130 for width, but got 30. 

<div class="container" data-expected-width="130" data-expected-height="130">
    <img class="min-content" src="resources/square-blue-100x100.png" style="width: 100%; border: 5px solid black; padding: 5px;" data-expected-width="120" data-expected-height="120">
</div>
FAIL:
Expected 130 for width, but got 30. 

<div class="container" data-expected-width="130" data-expected-height="130">
    <img class="max-content" src="resources/square-blue-100x100.png" style="width: 100%; border: 5px solid black; padding: 5px;" data-expected-width="120" data-expected-height="120">
</div>
PASS
PASS
PASS
PASS
PASS
